requirements.txt
mss>=9.0.1
Pillow>=10.0.0
PyQt5>=5.15.9
